Make Code an enum from int64, not int

+++ buse/busegen/utils.go
@@ -30,7 +30,7 @@ func isEnum(ts *ast.TypeSpec) bool {
 	}
 
 	if id, ok := ts.Type.(*ast.Ident); ok {
-		if id.Name == "string" || id.Name == "int" {
+		if id.Name == "string" || id.Name == "int64" {
 			return true
 		}
 	}
